🗃️ ftree: split auth module controllers into accouns and authorization
